renderer: C++ compatibility for tr_main

+++ code/renderer/tr_main.c
@@ -1494,12 +1494,12 @@ static void R_SortDrawSurfs( drawSurf_t *drawSurfs, int numDrawSurfs ) {
 	for ( i = 0 ; i < numDrawSurfs ; i++ ) {
 		R_DecomposeSort( (drawSurfs+i)->sort, &entityNum, &shader, &fogNum, &dlighted );
 
-		if ( shader->sort > SS_PORTAL ) {
+		if ( shader->sort > (float)SS_PORTAL ) {
 			break;
 		}
 
 		// no shader should ever have this sort type
-		if ( shader->sort == SS_BAD ) {
+		if ( shader->sort == (float)SS_BAD ) {
 			ri.Error (ERR_DROP, "Shader '%s'with sort == SS_BAD", shader->name );
 		}
